25/11/2022

+++ database/migrations/2022_10_02_170339_create_provinsi_table.php
@@ -14,7 +14,7 @@ class CreateProvinsiTable extends Migration
     public function up()
     {
         Schema::create('provinsi', function (Blueprint $table) {
-            $table->unsignedInteger('id')->primary();
+            $table->unsignedBigInteger('id')->primary();
             $table->string('nama')->nullable();
         });
     }
